refactor pm quality review action detail router
