+++ src/app/router_registry.py
@@ -113,6 +113,9 @@ from app.routers.dpm_command_center_pm_quality_policies import (
 from app.routers.dpm_command_center_pm_quality_policy_actions import (
     router as dpm_command_center_pm_quality_policy_actions_router,
 )
+from app.routers.dpm_command_center_pm_quality_review_action_detail import (
+    router as dpm_command_center_pm_quality_review_action_detail_router,
+)
 from app.routers.dpm_command_center_pm_quality_review_action_lookup import (
     router as dpm_command_center_pm_quality_review_action_lookup_router,
 )
@@ -389,6 +392,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_pm_quality_policies_router)
     app.include_router(dpm_command_center_pm_quality_policy_actions_router)
     app.include_router(dpm_command_center_pm_quality_review_actions_router)
+    app.include_router(dpm_command_center_pm_quality_review_action_detail_router)
     app.include_router(dpm_command_center_pm_quality_review_action_lookup_router)
     app.include_router(dpm_command_center_pm_quality_score_run_lookup_router)
     app.include_router(dpm_command_center_pm_quality_summary_invocations_router)
